+++ src/Gate.IO.Api/Account/Requests/GateAccountDebitFeeRequest.cs
@@ -1,0 +1,12 @@
+namespace Gate.IO.Api.Account;
+
+/// <summary>
+/// Account GT fee deduction request
+/// </summary>
+public record GateAccountDebitFeeRequest
+{
+    /// <summary>
+    /// Whether GT fee deduction is enabled
+    /// </summary>
+    public bool Enabled { get; set; }
+}