+++ src/Gate.IO.Api/Account/Requests/GateAccountStpGroupUsersRequest.cs
@@ -1,0 +1,12 @@
+namespace Gate.IO.Api.Account;
+
+/// <summary>
+/// Account STP group users request
+/// </summary>
+public record GateAccountStpGroupUsersRequest
+{
+    /// <summary>
+    /// User IDs
+    /// </summary>
+    public IEnumerable<long> UserIds { get; set; }
+}