+++ src/Gate.IO.Api/Account/Responses/GateAccountStpGroup.cs
@@ -9,7 +9,14 @@ public record GateAccountStpGroup
     public string Name { get; set; }
 
     [JsonProperty("creator_id")]
-    public long CreateId { get; set; }
+    public long CreatorId { get; set; }
+
+    [JsonIgnore]
+    public long CreateId
+    {
+        get => CreatorId;
+        set => CreatorId = value;
+    }
 
     [JsonProperty("create_time")]
     [JsonConverter(typeof(DateTimeConverter))]